Use appContext to only play one audio at a time
