+++ content/webapp/components/AudioPlayerNew/index.tsx
@@ -1,6 +1,5 @@
 import * as prismic from '@prismicio/client';
 import { FunctionComponent, useEffect, useRef, useState } from 'react';
-import styled, { css } from 'styled-components';
 
 import { useAppContext } from '@weco/common/contexts/AppContext';
 import { font } from '@weco/common/utils/classnames';
@@ -18,130 +17,18 @@ import {
 } from './AudioPlayer.Icons';
 import PlayRate from './AudioPlayer.PlayRate';
 import Scrubber from './AudioPlayer.Scrubber';
-
-const AudioPlayerWrapper = styled(Space).attrs({
-  as: 'figure',
-  $v: { size: 'm', properties: ['padding-top', 'padding-bottom'] },
-})<{ $isDark: boolean }>`
-  background: ${props =>
-    props.$isDark ? props.theme.color('black') : props.theme.color('white')};
-  margin: 0;
-`;
-
-type PlayPauseButtonProps = { $isPlaying: boolean };
-const PlayPauseButton = styled.button.attrs<PlayPauseButtonProps>(props => ({
-  'aria-pressed': props.$isPlaying,
-}))<PlayPauseButtonProps>`
-  padding: ${props => props.theme.spacingUnits['5']}px
-    ${props => props.theme.spacingUnits['6']}px 0;
-`;
-
-const TimeWrapper = styled.div.attrs({
-  className: font('intr', 6),
-})<{ $isDark: boolean }>`
-  display: flex;
-  justify-content: space-between;
-  font-variant-numeric: tabular-nums;
-  color: ${props =>
-    props.$isDark ? props.theme.color('white') : props.theme.color('black')};
-`;
-
-const SkipPlayWrapper = styled.div`
-  grid-column: 2;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-`;
-
-const colorTransform = css<{ $isDark: boolean }>`
-  color: ${props =>
-    props.$isDark ? props.theme.color('yellow') : props.theme.color('black')};
-  transform: scale(1.1);
-`;
-
-const SkipButton = styled.button<{ $isDark: boolean }>`
-  padding: ${props => props.theme.spacingUnits['5']}px 0 0;
-  color: ${props =>
-    props.$isDark ? props.theme.color('white') : props.theme.color('black')};
-
-  transition:
-    color 0.2s ease-out,
-    transform 0.2s ease-out;
-
-  @media (hover: hover) {
-    &:hover {
-      ${colorTransform};
-    }
-  }
-
-  @media (hover: none) {
-    &:active {
-      ${colorTransform};
-    }
-  }
-`;
-
-const PlayerRateWrapper = styled.div`
-  padding-top: ${props => props.theme.spacingUnits['5']}px;
-  grid-column: 3;
-  display: flex;
-  align-items: center;
-  justify-content: end;
-`;
-
-const colorTransformIconFill = css<{ $isDark: boolean }>`
-  color: ${props =>
-    props.$isDark ? props.theme.color('yellow') : props.theme.color('black')};
-  transform: scale(1.1);
-
-  .icon__playpause {
-    fill: ${props =>
-      props.$isDark ? props.theme.color('black') : props.theme.color('white')};
-  }
-`;
-
-const TitleWrapper = styled.span<{ $isDark: boolean }>`
-  color: ${props =>
-    props.$isDark ? props.theme.color('white') : props.theme.color('black')};
-`;
-
-const PlayPauseInner = styled.div<{ $isDark: boolean }>`
-  color: ${props =>
-    props.$isDark ? props.theme.color('white') : props.theme.color('black')};
-  transition:
-    color 0.2s ease-out,
-    transform 0.2s ease-out;
-
-  .icon__playpause {
-    fill: ${props =>
-      props.$isDark ? props.theme.color('black') : props.theme.color('white')};
-    transition: fill 0.2s ease-out;
-  }
-
-  @media (hover: hover) {
-    &:hover {
-      ${colorTransformIconFill};
-    }
-  }
-
-  @media (hover: none) {
-    &:active {
-      ${colorTransformIconFill};
-    }
-  }
-`;
-
-const AudioPlayerGrid = styled.div.attrs({})<{ $isEnhanced: boolean }>`
-  display: ${props => (props.$isEnhanced ? 'grid' : 'none')};
-  grid-template-columns: repeat(3, 1fr);
-  align-items: center;
-`;
-
-const NowPlayingWrapper = styled(Space).attrs({
-  $v: { size: 's', properties: ['margin-top'] },
-})`
-  grid-column: 1 / -1;
-`;
+import {
+  AudioPlayerGrid,
+  AudioPlayerWrapper,
+  NowPlayingWrapper,
+  PlayerRateWrapper,
+  PlayPauseButton,
+  PlayPauseInner,
+  SkipButton,
+  SkipPlayWrapper,
+  TimeWrapper,
+  TitleWrapper,
+} from './AudioPlayer.styles';
 
 export type AudioPlayerProps = {
   audioFile: string;
@@ -156,7 +43,7 @@ export const AudioPlayer: FunctionComponent<AudioPlayerProps> = ({
   title,
   isDark,
   transcript,
-  titleProps = {},
+  titleProps,
 }) => {
   const { isEnhanced } = useAppContext();
   const [currentTime, setCurrentTime] = useState(0);
@@ -168,6 +55,7 @@ export const AudioPlayer: FunctionComponent<AudioPlayerProps> = ({
   // announcement every second.
   const [startTime, setStartTime] = useState(currentTime);
   const { trackPlay, trackEnded, trackTimeUpdate } = useAVTracking('audio');
+  const { activeAudioPlayerId, setActiveAudioPlayerId } = useAppContext();
 
   const audioPlayerRef = useRef<HTMLAudioElement>(null);
   const progressBarRef = useRef<HTMLInputElement>(null);
@@ -228,6 +116,13 @@ export const AudioPlayer: FunctionComponent<AudioPlayerProps> = ({
       audioPlayerRef.current.play();
     }
   };
+
+  // Pause playing if another audio player is active
+  useEffect(() => {
+    if (activeAudioPlayerId !== audioFile && isPlaying) {
+      onTogglePlay();
+    }
+  }, [activeAudioPlayerId]);
 
   const handleSkipBackClick = () => {
     if (!audioPlayerRef.current) return;
@@ -313,12 +208,7 @@ export const AudioPlayer: FunctionComponent<AudioPlayerProps> = ({
                   </span>
                 )}
               </TimeWrapper>
-              <Space
-                $v={{
-                  size: 's',
-                  properties: ['padding-top'],
-                }}
-              >
+              <Space $v={{ size: 's', properties: ['padding-top'] }}>
                 <Space $v={{ size: 'xs', properties: ['padding-bottom'] }}>
                   <Scrubber
                     startTime={startTime}
@@ -333,6 +223,7 @@ export const AudioPlayer: FunctionComponent<AudioPlayerProps> = ({
               </Space>
             </div>
           </NowPlayingWrapper>
+
           <SkipPlayWrapper>
             <SkipButton $isDark={!!isDark} onClick={handleSkipBackClick}>
               <span className="visually-hidden">rewind 15 seconds</span>
@@ -352,6 +243,7 @@ export const AudioPlayer: FunctionComponent<AudioPlayerProps> = ({
               <SkipForwardIcon />
             </SkipButton>
           </SkipPlayWrapper>
+
           {audioPlayerRef.current && (
             <PlayerRateWrapper>
               <PlayRate
@@ -369,6 +261,8 @@ export const AudioPlayer: FunctionComponent<AudioPlayerProps> = ({
           onPlay={event => {
             trackPlay(event);
             setIsPlaying(true);
+            if (activeAudioPlayerId !== audioFile)
+              setActiveAudioPlayerId(audioFile);
           }}
           onEnded={trackEnded}
           onPause={() => setIsPlaying(false)}
